Fix an issue where switching b/w engrams while editor is focused causes a POST request
This issue is also why rapidly switching b/w engrams while making the POST update request can cause blocks to be deleted (the editor is focused and swapping the blocks makes it think blocks are deleted by the user)

By adding the key attribute this way, the Editor component is forced to re-render when going from one engram to the other. This way, there's no reason to worry about the focus editor state persisting, because it will be reset.